+++ sdk/highlight-next/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @highlight-run/next
 
+## 7.5.10
+
+### Patch Changes
+
+-   Updated dependencies [50c74161e]
+    -   highlight.run@9.0.3
+    -   @highlight-run/node@3.9.0
+
+## 7.5.9
+
+### Patch Changes
+
+-   Updated dependencies [fd011e719]
+-   Updated dependencies [1dbe1bc6c]
+    -   highlight.run@9.0.2
+    -   @highlight-run/react@3.2.3
+    -   @highlight-run/node@3.9.0
+
 ## 7.5.8
 
 ### Patch Changes